regeared the code base

+++ lib/features/introduction_page/Interface/introduction_page.dart
@@ -135,7 +135,7 @@ class _IntroductionPageState extends State<IntroductionPage> {
                   child: Center(
                     child: DotsIndicator(
                       dotsCount: dotsLength,
-                      position: currentDot,
+                      position: (currentDot).toDouble(),
                       //* Limit to original items
                       decorator: DotsDecorator(
                         shape: CircleBorder(
